# packages/api-client/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/sessions": {
      "get": {
        "operationId": "AuthController_sessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/devices": {
      "post": {
        "operationId": "AuthController_registerDevice",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/sessions/{id}/revoke": {
      "post": {
        "operationId": "AuthController_revokeSession",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "operationId": "AuthController_logoutAll",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/{entity}": {
      "get": {
        "operationId": "ReferenceController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "references"
        ]
      },
      "post": {
        "operationId": "ReferenceController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "references"
        ]
      }
    },
    "/api/v1/users": {
      "post": {
        "operationId": "ReferenceController_createUser",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "references"
        ]
      }
    },
    "/api/v1/users/{id}/roles": {
      "post": {
        "operationId": "ReferenceController_assignRoles",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "references"
        ]
      }
    },
    "/api/v1/users/{id}/block": {
      "post": {
        "operationId": "ReferenceController_setUserBlocked",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "references"
        ]
      }
    },
    "/api/v1/settings": {
      "post": {
        "operationId": "ReferenceController_upsertSetting",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "references"
        ]
      }
    },
    "/api/v1/{entity}/{id}": {
      "patch": {
        "operationId": "ReferenceController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "references"
        ]
      }
    },
...
    "/api/v1/waybills/{id}": {
      "get": {
        "operationId": "WaybillsController_getById",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "waybills"
        ]
      }
    },
...
    "/api/v1/washing-batches": {
      "post": {
        "operationId": "OperationsController_washing",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "operations"
        ]
      },
      "get": {
        "operationId": "OperationsController_listWashing",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/production-batches": {
      "post": {
        "operationId": "OperationsController_production",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "operations"
        ]
      },
      "get": {
        "operationId": "OperationsController_listProduction",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "operations"
        ]
      }
    },
...
    "/api/v1/files/upload-url": {
      "post": {
        "operationId": "FilesController_createUploadUrl",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "files"
        ]
      }
    },
    "/api/v1/files/{id}/download-url": {
      "get": {
        "operationId": "FilesController_createDownloadUrl",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "files"
        ]
      }
    },
    "/api/v1/files/waybills/{waybillId}": {
      "post": {
        "operationId": "FilesController_attachToWaybill",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "files"
        ]
      }
    },
...
  },
  "info": {
    "title": "BIOFLOW API",
    "description": "Control of extraction, transport, acceptance, inventory, washing and production.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}